moved tenzies dice game into /work page
