+++ src/app/work/tenzies-dice-game/page.tsx
@@ -1,0 +1,24 @@
+import { draftMode } from "next/headers";
+import { client, previewClient } from "@/src/lib/client";
+import { PageInterface } from "@/types/PageInterface";
+import BaseTemplate from "@/src/templates/BaseTemplate";
+import Canvas from "./Canvas";
+
+export default async function Page() {
+  const { isEnabled: preview } = draftMode();
+  const gqlClient = preview ? previewClient : client;
+
+  const landingPageData = await gqlClient.Page({
+    slug: "/work/tenzies-dice-game",
+    preview,
+    locale: "en-US",
+  });
+  // @ts-ignore
+  const page: PageInterface = landingPageData.pageCollection?.items[0];
+
+  return (
+    <BaseTemplate {...page}>
+      <Canvas />
+    </BaseTemplate>
+  );
+}
